create module and move merging functions
